upgrade/scripts/requirements.txt
# Add your required packages below
azure-identity
azure-mgmt-compute
azure-mgmt-subscription
azure.mgmt.resource
pyyaml
inquirer
